+++ packages/recoil/src/atoms/solana/index.tsx
@@ -1,4 +1,5 @@
 export * from "./jupiter";
+export * from "./nft";
 export * from "./preferences";
 export * from "./recent-transactions";
 export * from "./token";
